dedupe: Stop dedupe deleting files with identical IDs #4013
Before this change if there were two files with the same name and the
same ID in the same directory, dedupe would delete one of them but
since these are actually the same file (with the same ID) then both
files would be deleted leading to data loss.

This situation is quite hard to create so hasn't come up until now.

This change checks to see if any of the duplicates have the same ID
and if they do it refuses to delete them.